# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hikāya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    get:
      tags: [stories]
      operationId: listStories
      summary: List stories
      parameters:
        - name: language
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: type
          in: query
          required: false
          schema: { type: string }
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'
  /stories/{slug}:
    get:
      tags: [stories]
      operationId: getStoryBySlug
      summary: Get story by slug with cast and chapter list
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /stories/{slug}/chapters/{chapterNumber}:
    get:
      tags: [stories]
      operationId: getChapter
      summary: Get a chapter
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: chapterNumber
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Story:
      type: object
      required: [id, slug, title, description, language, category, type, tags, status, accentColor, createdAt]
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        coverImage: { type: [string, 'null'] }
        language: { type: string }
        category: { type: string }
        type: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
        videoUrl: { type: [string, 'null'] }
        accentColor: { type: string }
        createdAt: { type: string, format: date-time }
        chapterCount: { type: integer }
    Character:
      type: object
      required: [id, storyId, name, role, voiceId, tone, color, stability, similarity, style]
      properties:
        id: { type: string }
        storyId: { type: string }
        name: { type: string }
        role: { type: string }
        voiceId: { type: string }
        tone: { type: string }
        color: { type: string }
        stability: { type: integer }
        similarity: { type: integer }
        style: { type: integer }
    Chapter:
      type: object
      required: [id, storyId, chapterNumber, title, content, segments, durationSec]
      properties:
        id: { type: string }
        storyId: { type: string }
        chapterNumber: { type: integer }
        title: { type: string }
        content: { type: string }
        segments:
          type: array
          items: { $ref: '#/components/schemas/ScriptSegment' }
        durationSec: { type: integer }
        hasAudio: { type: boolean }
    ChapterSummary:
      type: object
      required: [id, chapterNumber, title, durationSec, hasAudio]
      properties:
        id: { type: string }
        chapterNumber: { type: integer }
        title: { type: string }
        durationSec: { type: integer }
        hasAudio: { type: boolean }
    StoryDetail:
      type: object
      required: [story, characters, chapters]
      properties:
        story:
          $ref: '#/components/schemas/Story'
        characters:
          type: array
          items: { $ref: '#/components/schemas/Character' }
        chapters:
          type: array
          items: { $ref: '#/components/schemas/ChapterSummary' }
        bookmarked: { type: boolean }
        progress:
          oneOf:
            - $ref: '#/components/schemas/ProgressEntry'
            - type: 'null'
    ChapterDetail:
      type: object
      required: [story, chapter, characters]
      properties:
        story: { $ref: '#/components/schemas/Story' }
        chapter: { $ref: '#/components/schemas/Chapter' }
        characters:
          type: array
          items: { $ref: '#/components/schemas/Character' }
        nextChapterNumber: { type: [integer, 'null'] }
        prevChapterNumber: { type: [integer, 'null'] }
    ScriptSegment:
      type: object
      required: [index, kind, characterName, emotion, text]
      properties:
        index: { type: integer }
        kind:
          type: string
          enum: [narrator, dialogue, scene]
        characterId: { type: [string, 'null'] }
        characterName: { type: string }
        emotion: { type: string }
        text: { type: string }
        audioUrl: { type: [string, 'null'] }
    ProgressEntry:
      type: object
      required: [storyId, chapterNumber, progressPercent, updatedAt, story]
      properties:
        storyId: { type: string }
        chapterNumber: { type: integer }
        progressPercent: { type: integer }
        updatedAt: { type: string, format: date-time }
        story: { $ref: '#/components/schemas/Story' }
...