# lib/api-zod/src/generated/types/chapterDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
import type { Chapter } from "./chapter";
import type { Character } from "./character";
import type { Story } from "./story";

export interface ChapterDetail {
  story: Story;
  chapter: Chapter;
  characters: Character[];
  /** @nullable */
  nextChapterNumber?: number | null;
  /** @nullable */
  prevChapterNumber?: number | null;
}