# lib/api-zod/src/generated/types/chapterSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */

export interface ChapterSummary {
  id: string;
  chapterNumber: number;
  title: string;
  durationSec: number;
  hasAudio: boolean;
}