# lib/api-zod/src/generated/types/storyDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterSummary } from "./chapterSummary";
import type { Character } from "./character";
import type { ProgressEntry } from "./progressEntry";
import type { Story } from "./story";

export interface StoryDetail {
  story: Story;
  characters: Character[];
  chapters: ChapterSummary[];
  bookmarked?: boolean;
  progress?: ProgressEntry | null;
}